drm/exynos: add host_ops callback for platform drivers
Platform drivers need to be aware if a mipi dsi device attaches or
detaches. Add host_ops to the driver_data for attach and detach
callbacks and move the i80 mode selection and the hotplug handling into
the callback, because these depend on the drm driver.

Upstream-Status: Posted [https://patchwork.freedesktop.org/patch/390176/?series=81295&rev=2]
Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>