drm/exynos: configure mode on drm bridge
The driver uses the encoder to get the mode that shall be configured.
This is not possible, if the driver is used as a bridge, because the
encoder might not be used.

Use the mode_set function to set the display mode.

Upstream-Status: Posted [https://patchwork.freedesktop.org/patch/390180/?series=81295&rev=2]
Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>