firmware: arm_scmi: Fix pinctrl fwdevlink
We have two drivers, pinctrl-scmi.c which is generic and
pinctrl-imx-scmi.c which is for IMX hardware. They do the same
thing.  Both provide support for the SCMI_PROTOCOL_PINCTRL protocol.

If you have a kernel with both modules built in then they way it
was designed to work is that the probe() functions would only
allow the appropriate driver to probe.  Unfortunately, what happens
is that there is an issue earlier in the process so the
fwnode device points to the wrong driver. This means that
even though the correct driver is probed it still wants to use
whichever driver was loaded first so if the driver you want came
second then it won't work.

The other issue is:
Vendor A use 0x80 for feature X, Vendor B use 0x80 for feature Y.
With both drivers built in, two scmi devices will be created, and both
drivers will be probed. On A's patform, feature Y probe may fail, vice
verus.

To fix this, introduce allowed_ids and blocked_ids for scmi_device_id.
Only create the scmi device when allowed in scmi_create_device() function.

Fixes: b755521fd6eb ("pinctrl: imx: support SCMI pinctrl protocol for i.MX95")
Cc: Dan Carpenter <dan.carpenter@linaro.org>
Signed-off-by: Peng Fan <peng.fan@nxp.com>